// examples/optimistic-demo/convex/_generated/dataModel.d.ts
// Generated by Stackbase. Do not edit.
import type { GenericId } from "@stackbase/values";

export interface DataModel {
  _storage: { document: { _id: Id<"_storage">; _creationTime: number; status: ("pending" | "ready"); key: string; size: (number | null); contentType: (string | null); sha256: (string | null); visibility: ("private" | "public"); expiresAt: (number | null) } };
  polls: { document: { _id: Id<"polls">; _creationTime: number; question: string; closed: boolean } };
  options: { document: { _id: Id<"options">; _creationTime: number; pollId: Id<"polls">; label: string; votes: number; order: number } };
}

export type TableNames = keyof DataModel;
export type Id<TableName extends TableNames> = GenericId<TableName>;
export type Doc<TableName extends TableNames> = DataModel[TableName]["document"];